+++ src/components/Korea/Korea.jsx
@@ -1,6 +1,9 @@
+import LoginMap from '../Maps/LoginMap';
+
 function Korea() {
   return (
     <>
+      <LoginMap />
       <div>This is korea!</div>
     </>
   );
