+++ src/components/Maps/NonLoginMap.jsx
@@ -1,10 +1,11 @@
-import { useState, useCallback, memo, useEffect } from 'react';
-import { GoogleMap, useJsApiLoader, MarkerF, InfoWindow, InfoWindowF } from '@react-google-maps/api';
-import CreateMarkerForm from './CreateMarkerForm';
-import MarkerInfo from './MarkerInfo';
+import { useState, useCallback, memo } from 'react';
+import { GoogleMap, useJsApiLoader, MarkerF } from '@react-google-maps/api';
 import './Map.css';
 
-let center = { lat: 37.49, lng: 127.02 };
+const center = {
+  lat: -3.745,
+  lng: -38.523,
+};
 
 const myStyles = [
   {
@@ -20,10 +21,6 @@ const myStyles = [
 
 function NonLoginMap() {
   const [markers, setMarkers] = useState([]);
-  const [map, setMap] = useState(null);
-  const [userLocation, setUserLocation] = useState(null); // 현재 위치를 받아올 상태
-  const [creatingMarker, setCreatingMarker] = useState(false); // 마커를 만드는 중인지
-  const [selectedMarker, setSelectedMarker] = useState(null); // 저장된 마커의 정보를 보여줄 마커
 
   const { isLoaded } = useJsApiLoader({
     id: 'google-map-script',
@@ -33,109 +30,56 @@ function NonLoginMap() {
     version: 'quarterly',
   });
 
-  const onLoad = useCallback(
-    function callback(map) {
-      setMap(map);
+  const [map, setMap] = useState(null);
 
-      if (navigator.geolocation) {
-        navigator.geolocation.getCurrentPosition((position) => {
-          // center의 좌표를 현재 위치로
-          const lat = position.coords.latitude;
-          const lng = position.coords.longitude;
-          setUserLocation({ lat, lng });
-        });
-      }
-    },
-    [userLocation]
-  );
+  const onLoad = useCallback(function callback(map) {
+    const bounds = new window.google.maps.LatLngBounds(center);
+    map.fitBounds(bounds);
+
+    setMap(map);
+  }, []);
 
   const onUnmount = useCallback(function callback(map) {
     setMap(null);
   }, []);
 
   const handleMapClick = (event) => {
-    if (creatingMarker) {
-      if (markers.length > 0) {
-        setMarkers(markers.slice(0, markers.length - 1));
-      }
-      setCreatingMarker(false);
-    } else {
-      const lat = event.latLng.lat();
-      const lng = event.latLng.lng();
-
-      const newMarker = {
-        id: lat,
-        position: {
-          lat: lat,
-          lng: lng,
-        },
-      };
-
-      setCreatingMarker(true);
-      setMarkers([...markers, newMarker]);
-
-      if (map) {
-        map.panTo(newMarker.position);
-        console.log(newMarker);
-      }
-    }
-    // 마커 정보 표시 안함
-    setSelectedMarker(null);
-  };
-
-  const handleCreateMarker = (newMarkerInfo) => {
-    const lat = markers[markers.length - 1].position.lat;
-    const lng = markers[markers.length - 1].position.lng;
+    const lat = event.latLng.lat();
+    const lng = event.latLng.lng();
 
     const newMarker = {
-      id: lat,
-      position: { lat, lng },
-      info: { ...newMarkerInfo },
+      lat: lat,
+      lng: lng,
     };
 
-    setMarkers([...markers.slice(0, markers.length - 1), newMarker]);
-    setCreatingMarker(false);
-  };
-
-  const handleMarkerClick = (marker) => {
-    if (creatingMarker) {
-      // 생성 중인 마커를 클릭하면 해당 마커를 삭제
-      if (markers.length > 0) {
-        setMarkers(markers.slice(0, markers.length - 1));
-      }
-      setCreatingMarker(false);
-    }
-    setSelectedMarker(marker);
-    map.panTo(marker.position);
-    console.log(marker);
+    setMarkers([...markers, newMarker]);
+    // 원하는 작업 수행
+    // 예: 마커를 추가하거나 위치를 기록
   };
 
   return isLoaded ? (
-    <div className='map-display'>
-      <div>지도화면입니다</div>
+    <>
+      <div>This is Blank</div>
       <GoogleMap
         mapContainerClassName='map-container'
-        center={userLocation || center}
+        center={center}
+        zoom={10}
         onLoad={onLoad}
-        zoom={16}
         onUnmount={onUnmount}
         options={{ disableDefaultUI: true, styles: myStyles }}
         onClick={(event) => handleMapClick(event)}
       >
-        {markers.map((marker) => (
+        {markers.map((marker, index) => (
           <MarkerF
-            key={marker.id}
-            position={marker.position}
-            onClick={() => handleMarkerClick(marker)}
+            key={index}
+            position={marker}
             // label={(index + 1).toString()} // 마커에 숫자 레이블 추가
           />
         ))}
       </GoogleMap>
-      {creatingMarker && <CreateMarkerForm onCreateMarker={handleCreateMarker} />}
-      {selectedMarker && <MarkerInfo marker={selectedMarker} onClose={() => setSelectedMarker(null)} />}
-    </div>
+    </>
   ) : (
-    <div></div>
+    <></>
   );
 }
